Add files via upload
The evaluation brings out some performance trends correlating to the students' performance. The education attainment of the parents, race, and taking a test preparation course have been noticed to predict student performance. The higher the parents’ educational attainment, for instance, the presence of a master’s degree, the better the performance. Test scores also vary widely among the children, where the group that did the worst, group a scored the lowest and group c the highest. It is evident from the study that taking up test preparation positively influences student performance as evidenced by those who completed the course. The students, whose scores were comparatively lower had difficulty preparing for a test and completed an essay than those who did not.